# Remove RTree Delete and Insert methods