… rtree/delete.go
… rtree/insert.go