## Description

There are a few reasons why these are removed:

- GEOS' `STRtree` type is a "bulk only" r-tree, disallowing entries to be added or deleted after the first query to the tree.

- Simplefeatures itself doesn't use the `Insert` or `Delete` methods.

- The `Insert` and `Delete` methods are _very_ complicated.

- The `Insert` and `Delete` methods are much slower than the equivalent bulk-load operation.

- The presence of the `Insert` and `Delete` methods add extra constraints to the tree structure, that make it difficult to modify and improve the bulk-loading algorithms.

- There already exists other very high quality r-tree implementations that can perform insert and delete, which gives existing users of the simplefeatures `Insert` and `Delete` methods a different package to migrate to if needed. For example, `https://github.com/tidwall/rtree`.

## Check List

Have you:

- Added unit tests? N/A

- Add cmprefimpl tests? (if appropriate?) N/A

- Updated release notes? (if appropriate?) Yes.

## Related Issue

- https://github.com/peterstace/simplefeatures/issues/483

## Benchmarks

<details>

<summary>Click to open</summary>

```
goos: linux
goarch: arm64
pkg: github.com/peterstace/simplefeatures/geom
                                                           │ /tmp/tmp.4Q3TPX7f8Q │         /tmp/tmp.WA74bwKl4o          │
                                                           │       sec/op        │    sec/op     vs base                │
LineEnvelope/0-2                                                   2.155n ±   5%   2.140n ±  2%        ~ (p=0.418 n=15)
LineEnvelope/1-2                                                   2.130n ±   1%   2.119n ±  1%        ~ (p=0.615 n=15)
LineEnvelope/2-2                                                   2.094n ±   1%   2.093n ±  0%        ~ (p=0.587 n=15)
LineEnvelope/3-2                                                   2.100n ±   0%   2.100n ±  0%        ~ (p=0.505 n=15)
MarshalWKB/polygon/n=10-2                                          92.51n ±   4%   92.77n ±  1%        ~ (p=0.806 n=15)
MarshalWKB/polygon/n=100-2                                         236.6n ±   8%   236.1n ±  2%        ~ (p=0.870 n=15)
MarshalWKB/polygon/n=1000-2                                        1.611µ ±   5%   1.651µ ±  9%        ~ (p=0.281 n=15)
MarshalWKB/polygon/n=10000-2                                       37.67µ ±  28%   41.90µ ± 19%        ~ (p=0.713 n=15)
UnmarshalWKB/polygon/n=10-2                                        156.9n ±   4%   157.8n ±  3%        ~ (p=0.430 n=15)
UnmarshalWKB/polygon/n=100-2                                       303.5n ±   4%   297.4n ±  5%        ~ (p=0.705 n=15)
UnmarshalWKB/polygon/n=1000-2                                      1.836µ ±   8%   1.823µ ± 16%        ~ (p=0.926 n=15)
UnmarshalWKB/polygon/n=10000-2                                     38.00µ ±  18%   45.16µ ± 23%        ~ (p=0.250 n=15)
IntersectsLineStringWithLineString/n=10-2                          782.6n ±   1%   731.0n ±  1%   -6.59% (p=0.000 n=15)
IntersectsLineStringWithLineString/n=100-2                         11.87µ ±   1%   11.43µ ±  1%   -3.74% (p=0.000 n=15)
IntersectsLineStringWithLineString/n=1000-2                        143.4µ ±   8%   139.5µ ± 14%        ~ (p=0.967 n=15)
IntersectsLineStringWithLineString/n=10000-2                       3.420m ±  15%   2.086m ±  8%  -39.01% (p=0.000 n=15)
IntersectsMultiPointWithMultiPoint/n=20-2                          555.1n ±   4%   554.3n ±  0%        ~ (p=0.245 n=15)
IntersectsMultiPointWithMultiPoint/n=200-2                         7.541µ ±   0%   7.511µ ±  0%        ~ (p=0.066 n=15)
IntersectsMultiPointWithMultiPoint/n=2000-2                        71.13µ ±   0%   71.32µ ±  0%   +0.27% (p=0.037 n=15)
IntersectsMultiPointWithMultiPoint/n=20000-2                       780.3µ ±   2%   780.4µ ±  0%        ~ (p=0.567 n=15)
PolygonSingleRingValidation/n=10-2                                 1.294µ ±   2%   1.263µ ±  1%   -2.40% (p=0.000 n=15)
PolygonSingleRingValidation/n=100-2                                19.17µ ±   2%   18.82µ ±  0%   -1.85% (p=0.000 n=15)
PolygonSingleRingValidation/n=1000-2                               276.3µ ±   1%   273.4µ ±  1%   -1.07% (p=0.001 n=15)
PolygonSingleRingValidation/n=10000-2                              4.563m ±   8%   3.381m ±  1%  -25.90% (p=0.000 n=15)
PolygonMultipleRingsValidation/n=4-2                               3.663µ ±   2%   3.621µ ±  0%   -1.15% (p=0.001 n=15)
PolygonMultipleRingsValidation/n=36-2                              30.66µ ±   1%   30.37µ ±  1%   -0.95% (p=0.008 n=15)
PolygonMultipleRingsValidation/n=400-2                             405.2µ ±   1%   400.9µ ±  1%   -1.06% (p=0.026 n=15)
PolygonMultipleRingsValidation/n=4096-2                            5.180m ±   1%   5.007m ±  1%   -3.34% (p=0.000 n=15)
PolygonZigZagRingsValidation/n=10-2                                6.169µ ±   3%   6.006µ ±  0%   -2.64% (p=0.000 n=15)
PolygonZigZagRingsValidation/n=100-2                               74.01µ ±   4%   71.88µ ±  0%   -2.88% (p=0.000 n=15)
PolygonZigZagRingsValidation/n=1000-2                              972.9µ ±   1%   990.5µ ±  3%   +1.80% (p=0.009 n=15)
PolygonZigZagRingsValidation/n=10000-2                             16.01m ±  13%   12.56m ±  1%  -21.60% (p=0.000 n=15)
PolygonAnnulusValidation/n=10-2                                    1.975µ ±   1%   1.939µ ±  2%   -1.82% (p=0.001 n=15)
PolygonAnnulusValidation/n=100-2                                   19.12µ ±   1%   18.74µ ±  1%   -1.97% (p=0.000 n=15)
PolygonAnnulusValidation/n=1000-2                                  336.8µ ±   2%   330.7µ ±  1%   -1.81% (p=0.000 n=15)
PolygonAnnulusValidation/n=10000-2                                 4.550m ±   4%   4.257m ±  1%   -6.44% (p=0.000 n=15)
MultipolygonValidation/n=1-2                                       218.3n ±   2%   212.6n ±  1%   -2.61% (p=0.001 n=15)
MultipolygonValidation/n=4-2                                       488.8n ±   2%   483.5n ±  2%        ~ (p=0.072 n=15)
MultipolygonValidation/n=16-2                                      2.139µ ±   1%   2.094µ ±  2%   -2.10% (p=0.000 n=15)
MultipolygonValidation/n=64-2                                      10.34µ ±   2%   10.06µ ±  1%   -2.68% (p=0.003 n=15)
MultipolygonValidation/n=256-2                                     55.25µ ±   1%   56.37µ ±  1%        ~ (p=0.061 n=15)
MultipolygonValidation/n=1024-2                                    342.9µ ±   3%   343.1µ ±  4%        ~ (p=0.389 n=15)
MultiPolygonTwoCircles/n=10-2                                      1.941µ ±   2%   1.855µ ±  1%   -4.43% (p=0.000 n=15)
MultiPolygonTwoCircles/n=100-2                                     23.14µ ±   8%   22.14µ ±  1%   -4.33% (p=0.000 n=15)
MultiPolygonTwoCircles/n=1000-2                                    318.0µ ±   8%   270.2µ ±  3%  -15.03% (p=0.000 n=15)
MultiPolygonTwoCircles/n=10000-2                                   5.017m ±   9%   4.668m ±  6%   -6.96% (p=0.001 n=15)
MultiPolygonMultipleTouchingPoints/n=1-2                           2.612µ ±   1%   2.555µ ±  1%   -2.18% (p=0.002 n=15)
MultiPolygonMultipleTouchingPoints/n=10-2                          20.93µ ±   4%   20.69µ ±  3%        ~ (p=0.056 n=15)
MultiPolygonMultipleTouchingPoints/n=100-2                         264.9µ ±   3%   259.7µ ±  1%   -1.95% (p=0.001 n=15)
MultiPolygonMultipleTouchingPoints/n=1000-2                        3.289m ±   6%   3.307m ±  3%        ~ (p=0.539 n=15)
WKTParsing/point-2                                                 1.122µ ±   4%   1.135µ ±  1%        ~ (p=0.245 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=false-2                26.69µ ±   5%   26.34µ ±  2%   -1.33% (p=0.005 n=15)
DistancePolygonToPolygonOrdering/n=100_swap=true-2                 26.75µ ±   3%   26.23µ ±  1%   -1.93% (p=0.002 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=false-2               466.3µ ±   1%   466.1µ ±  2%        ~ (p=0.838 n=15)
DistancePolygonToPolygonOrdering/n=1000_swap=true-2                465.4µ ±   1%   464.8µ ±  2%        ~ (p=0.412 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-2          3.244µ ±   2%   3.204µ ±  0%   -1.23% (p=0.000 n=15)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-2           3.244µ ±   3%   3.197µ ±  1%   -1.45% (p=0.002 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-2         34.84µ ±   3%   34.36µ ±  3%   -1.38% (p=0.023 n=15)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-2          34.81µ ±   4%   34.16µ ±  0%   -1.88% (p=0.000 n=15)
MultiLineStringIsSimpleManyLineStrings/n=100-2                     27.04µ ±   8%   25.95µ ±  1%   -4.03% (p=0.000 n=15)
MultiLineStringIsSimpleManyLineStrings/n=1000-2                    330.6µ ±   6%   321.6µ ±  1%   -2.73% (p=0.000 n=15)
ForceCWandForceCCW/0-2                                             16.10n ±   4%   16.07n ±  1%        ~ (p=0.236 n=15)
ForceCWandForceCCW/0#01-2                                          103.0n ±   3%   102.7n ±  1%        ~ (p=0.461 n=15)
ForceCWandForceCCW/1-2                                             16.25n ±   1%   16.21n ±  0%        ~ (p=0.139 n=15)
ForceCWandForceCCW/1#01-2                                          103.0n ±   3%   103.6n ±  3%        ~ (p=0.271 n=15)
ForceCWandForceCCW/2-2                                             29.79n ± 434%   29.69n ±  1%        ~ (p=0.051 n=15)
ForceCWandForceCCW/2#01-2                                          158.8n ±  81%   159.0n ±  4%        ~ (p=0.395 n=15)
ForceCWandForceCCW/3-2                                             30.02n ±   1%   29.74n ±  1%   -0.93% (p=0.010 n=15)
ForceCWandForceCCW/3#01-2                                          160.2n ±   6%   159.4n ±  4%        ~ (p=0.367 n=15)
ForceCWandForceCCW/4-2                                             43.48n ±   7%   43.29n ±  1%        ~ (p=0.057 n=15)
ForceCWandForceCCW/4#01-2                                          245.1n ±   4%   245.7n ±  2%        ~ (p=0.705 n=15)
ForceCWandForceCCW/5-2                                             43.49n ±   2%   43.30n ±  2%        ~ (p=0.051 n=15)
ForceCWandForceCCW/5#01-2                                          246.6n ±   4%   246.3n ±  3%        ~ (p=0.560 n=15)
ForceCWandForceCCW/6-2                                             62.94n ±   1%   62.53n ±  1%   -0.65% (p=0.012 n=15)
ForceCWandForceCCW/6#01-2                                          403.6n ±   3%   394.0n ±  1%        ~ (p=0.091 n=15)
EnvelopeTransformXY-2                                              140.0n ±   3%   139.1n ±  0%   -0.64% (p=0.000 n=15)
WKBParse/0-2                                                       422.5n ±   2%   421.0n ±  1%        ~ (p=0.237 n=15)
WKBParse/1-2                                                       418.8n ±   2%   416.7n ±  0%        ~ (p=0.083 n=15)
WKBParse/2-2                                                       483.9n ±   3%   482.6n ±  1%        ~ (p=0.129 n=15)
WKBParse/3-2                                                       483.5n ±   4%   484.7n ±  2%        ~ (p=0.588 n=15)
WKBParse/4-2                                                       577.6n ±   2%   569.4n ±  0%        ~ (p=0.281 n=15)
WKBParse/5-2                                                       424.3n ±   3%   424.6n ±  0%        ~ (p=0.645 n=15)
WKBParse/6-2                                                       499.0n ±   1%   496.2n ±  0%   -0.56% (p=0.010 n=15)
WKBParse/7-2                                                       486.5n ±   1%   488.5n ±  1%        ~ (p=0.237 n=15)
WKBParse/8-2                                                       568.7n ±   2%   566.6n ±  0%        ~ (p=0.087 n=15)
WKBParse/9-2                                                       310.7n ±   2%   310.2n ±  0%        ~ (p=0.309 n=15)
WKBParse/10-2                                                      311.7n ±   2%   311.2n ±  1%        ~ (p=0.190 n=15)
WKBParse/11-2                                                      312.1n ±   1%   311.2n ±  1%        ~ (p=0.254 n=15)
WKBParse/12-2                                                      308.8n ±   1%   309.2n ±  1%        ~ (p=0.430 n=15)
WKBParse/13-2                                                      595.3n ±   1%   596.5n ±  2%        ~ (p=0.430 n=15)
WKBParse/14-2                                                      703.2n ±   1%   703.6n ±  1%        ~ (p=0.506 n=15)
WKBParse/15-2                                                      703.9n ±   0%   702.0n ±  5%        ~ (p=0.894 n=15)
WKBParse/16-2                                                      849.8n ±   1%   850.7n ±  1%        ~ (p=0.870 n=15)
WKBParse/17-2                                                      701.3n ±   1%   701.0n ±  1%        ~ (p=0.559 n=15)
WKBParse/18-2                                                      896.8n ±   3%   896.2n ±  2%        ~ (p=0.645 n=15)
WKBParse/19-2                                                      908.8n ±   3%   908.5n ±  0%        ~ (p=0.546 n=15)
WKBParse/20-2                                                      1.101µ ±   2%   1.104µ ±  0%        ~ (p=0.846 n=15)
WKBParse/21-2                                                      301.1n ±   1%   301.3n ±  0%        ~ (p=0.573 n=15)
WKBParse/22-2                                                      302.9n ±   0%   302.6n ±  0%        ~ (p=0.814 n=15)
WKBParse/23-2                                                      302.1n ±   1%   303.1n ±  0%        ~ (p=0.519 n=15)
WKBParse/24-2                                                      298.4n ±   1%   299.0n ±  0%        ~ (p=0.170 n=15)
WKBParse/25-2                                                      1.564µ ±   0%   1.563µ ±  0%        ~ (p=0.492 n=15)
WKBParse/26-2                                                      1.979µ ±   1%   1.983µ ±  1%        ~ (p=0.894 n=15)
WKBParse/27-2                                                      1.979µ ±   0%   1.978µ ±  0%        ~ (p=0.894 n=15)
WKBParse/28-2                                                      2.476µ ±   2%   2.470µ ±  0%        ~ (p=0.170 n=15)
WKBParse/29-2                                                      300.3n ±   1%   300.4n ±  0%        ~ (p=0.894 n=15)
WKBParse/30-2                                                      302.5n ±   1%   302.6n ±  1%        ~ (p=0.602 n=15)
WKBParse/31-2                                                      303.2n ±   1%   302.0n ±  0%        ~ (p=0.350 n=15)
WKBParse/32-2                                                      297.9n ±   2%   297.9n ±  0%        ~ (p=0.705 n=15)
WKBParse/33-2                                                      572.2n ±   0%   570.4n ±  1%        ~ (p=0.690 n=15)
WKBParse/34-2                                                      651.2n ±   0%   652.8n ±  1%   +0.25% (p=0.049 n=15)
WKBParse/35-2                                                      650.2n ±   1%   654.3n ±  1%        ~ (p=0.395 n=15)
WKBParse/36-2                                                      724.6n ±   1%   725.1n ±  0%        ~ (p=0.690 n=15)
WKBParse/37-2                                                      819.6n ±   1%   820.1n ±  0%        ~ (p=0.351 n=15)
WKBParse/38-2                                                      971.5n ±   1%   980.8n ±  2%        ~ (p=0.113 n=15)
WKBParse/39-2                                                      968.7n ±   1%   971.5n ±  1%        ~ (p=0.202 n=15)
WKBParse/40-2                                                      1.079µ ±   1%   1.085µ ±  0%   +0.56% (p=0.033 n=15)
WKBParse/41-2                                                      301.9n ±   0%   301.9n ±  2%        ~ (p=0.862 n=15)
WKBParse/42-2                                                      304.8n ±   0%   304.3n ±  1%        ~ (p=0.418 n=15)
WKBParse/43-2                                                      303.0n ±   0%   303.4n ±  0%        ~ (p=0.532 n=15)
WKBParse/44-2                                                      298.9n ±   0%   298.6n ±  0%        ~ (p=0.228 n=15)
WKBParse/45-2                                                      888.9n ±   2%   886.9n ±  0%        ~ (p=0.113 n=15)
WKBParse/46-2                                                      1.079µ ±   1%   1.076µ ±  0%        ~ (p=0.087 n=15)
WKBParse/47-2                                                      1.087µ ±   1%   1.084µ ±  0%        ~ (p=0.068 n=15)
WKBParse/48-2                                                      1.242µ ±   0%   1.240µ ±  0%        ~ (p=0.307 n=15)
WKBParse/49-2                                                      1.273µ ±   0%   1.273µ ±  1%        ~ (p=0.992 n=15)
WKBParse/50-2                                                      1.607µ ±   1%   1.607µ ±  1%        ~ (p=0.910 n=15)
WKBParse/51-2                                                      1.603µ ±   1%   1.607µ ±  1%        ~ (p=0.767 n=15)
WKBParse/52-2                                                      1.874µ ±   0%   1.873µ ±  1%        ~ (p=0.862 n=15)
WKBParse/53-2                                                      301.8n ±   0%   301.9n ±  1%        ~ (p=0.588 n=15)
WKBParse/54-2                                                      303.8n ±   0%   303.7n ±  0%        ~ (p=0.798 n=15)
WKBParse/55-2                                                      303.2n ±   1%   304.0n ±  0%        ~ (p=0.720 n=15)
WKBParse/56-2                                                      298.8n ±   0%   299.3n ±  2%        ~ (p=0.309 n=15)
WKBParse/57-2                                                      1.927µ ±   2%   1.930µ ±  1%        ~ (p=0.660 n=15)
WKBParse/58-2                                                      2.361µ ±   0%   2.369µ ±  0%        ~ (p=0.244 n=15)
WKBParse/59-2                                                      2.371µ ±   2%   2.374µ ±  0%        ~ (p=0.505 n=15)
WKBParse/60-2                                                      2.872µ ±   1%   2.871µ ±  1%        ~ (p=0.736 n=15)
WKBParse/61-2                                                      301.2n ±   1%   300.5n ±  1%        ~ (p=0.134 n=15)
WKBParse/62-2                                                      304.0n ±   1%   302.6n ±  0%        ~ (p=0.072 n=15)
WKBParse/63-2                                                      303.5n ±   1%   302.2n ±  0%   -0.43% (p=0.012 n=15)
WKBParse/64-2                                                      298.7n ±   1%   297.7n ±  0%   -0.33% (p=0.044 n=15)
WKBParse/65-2                                                      875.9n ±   2%   875.7n ±  0%        ~ (p=0.735 n=15)
WKBParse/66-2                                                      1.020µ ±   2%   1.018µ ±  0%        ~ (p=0.075 n=15)
WKBParse/67-2                                                      1.033µ ±   5%   1.030µ ±  1%        ~ (p=0.190 n=15)
WKBParse/68-2                                                      1.144µ ±   2%   1.138µ ±  0%   -0.52% (p=0.009 n=15)
geomean                                                            1.949µ          1.924µ         -1.28%

                                                           │ /tmp/tmp.4Q3TPX7f8Q │          /tmp/tmp.WA74bwKl4o           │
                                                           │        B/op         │     B/op      vs base                  │
LineEnvelope/0-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/1-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/2-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
LineEnvelope/3-2                                                    0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=15) ¹
…